Add AST based parsing for varExtractor
Rename RegEx based varExtractor
Rename classes to match PEP8 guidelines
Untrack PyC files
